feat(chat): implement conversation thread and message components
- Added ConversationThread component for displaying chat messages.
- Introduced MessageBubble for individual message rendering with author info.
- Created ChatComposer for message input and sending functionality.
- Implemented GroupConversationAvatar for displaying group member avatars.
- Developed MessagesWorkspace to manage conversation lists and threads.
- Added NativeChatHeader for chat header with back navigation.
- Created RequestsBanner to show connection requests in chat.
- Established types for ChatMessage and Conversation for better type safety.

feat(feed): enhance feed components and post interactions

- Introduced FeedHeader for search functionality in the feed.
- Created FeedList to display a list of posts with load more functionality.
- Developed FeedPostCard for individual post display with reactions.
- Implemented PostThread for displaying post details and replies.
- Added NativeChatHeader for feed navigation with back button.
- Established types for FeedPost and GroupBoard for better type safety.
