+++ packages/frontend/components/chat/index.ts
@@ -1,0 +1,8 @@
+export { ConnectHeader } from './ConnectHeader'
+export { NativeChatHeader } from './NativeChatHeader'
+export { MessagesWorkspace } from './MessagesWorkspace'
+export { ConversationList } from './ConversationList'
+export { ConversationThread } from './ConversationThread'
+export { RequestsBanner } from './RequestsBanner'
+export { GroupConversationAvatar, GroupIcon } from './GroupAvatar'
+export type { ChatMessage, Conversation } from './types'
