+++ packages/frontend/components/connect/__mocks__/mockData.ts
@@ -175,71 +175,79 @@ const people = {
 
 export const eventBoards: EventBoard[] = [
   {
-    id: 'gita-12',
-    title: 'Bhagavad Gita Study · Ch. 12',
-    dateLabel: 'SAT, MAY 17',
-    centerLabel: 'Chinmaya San Jose',
-    attendeesLabel: '14 going',
-    preview: 'Anyone driving in from the Peninsula? Happy to coordinate a carpool.',
+    id: 'hanuman-chalisa-havan',
+    title: 'Hanuman Chalisa Havan',
+    dateLabel: 'SAT, MAY 16',
+    centerLabel: 'Chinmaya Portland',
+    attendeesLabel: '18 going',
+    preview: 'Anyone familiar with the Shodasha Upachara Puja sequence? First-timer here.',
     messages: [
       {
-        id: 'b1',
-        author: people.anjali,
-        timestamp: '2:14 PM',
-        body: 'Anyone driving in from the Peninsula? Happy to coordinate a carpool — I have room for 3.',
-        reactions: [{ emoji: '🚗', count: 4 }, { emoji: '🙏', count: 2 }],
-        replyCount: 2,
-      },
-      {
-        id: 'b2',
-        author: people.ravi,
-        timestamp: '4:02 PM',
-        body: "Quick reminder we're starting at 9:30 sharp this time, not 10. Aarti at 9:00 for those who can make it.",
-        reactions: [{ emoji: '🪔', count: 6 }],
-        replyCount: 1,
+        id: 'h1',
+        author: people.suresh,
+        timestamp: '8:00 AM',
+        body: 'Havan begins at 8:30 sharp. Please arrive by 8:00 for setup. Bring marigolds if you can.',
+        reactions: [{ emoji: '🪔', count: 9 }, { emoji: '🙏', count: 5 }],
+        replyCount: 3,
         pinned: true,
       },
       {
-        id: 'b3',
+        id: 'h2',
         author: people.priya,
-        timestamp: '8:11 AM',
-        body: "I'd love a ride if you've still got room — I'm in Burlingame.",
-        reactions: [{ emoji: '🙋', count: 3 }],
-        replyCount: 1,
+        timestamp: '9:22 AM',
+        body: "Anyone familiar with the Shodasha Upachara Puja sequence? Coming for the first time and want to follow along.",
+        reactions: [{ emoji: '🙋', count: 4 }],
+        replyCount: 2,
       },
       {
-        id: 'b4',
+        id: 'h3',
         author: people.karthik,
-        timestamp: '9:47 AM',
-        body: "Sharing a few photos from last week's session — hope folks who couldn't attend get a feel for it.",
-        attachmentLabel: 'Photo · seva session',
-        reactions: [{ emoji: '📸', count: 5 }],
-        replyCount: 3,
+        timestamp: '10:05 AM',
+        body: "I'll have a printed cheat-sheet of the Hanuman Chalisa verses for anyone who needs it — just ask.",
+        reactions: [{ emoji: '📄', count: 6 }],
+        replyCount: 1,
       },
     ],
   },
   {
-    id: 'youth-seva',
-    title: 'Youth Seva Day',
-    dateLabel: 'SUN, MAY 25',
-    centerLabel: 'Chinmaya Mission Tri-State',
-    attendeesLabel: '22 going',
-    preview: 'Can someone bring extra gloves and water coolers?',
+    id: 'chyk-memorial-day-camp',
+    title: 'ChYK Memorial Day Camp',
+    dateLabel: 'SAT, MAY 23 – MON, MAY 25',
+    centerLabel: 'ChYK West Central Zone · Chicago',
+    attendeesLabel: '31 going',
+    preview: 'Packing list is finalized — layers for evening, rain jacket just in case.',
     messages: [
       {
-        id: 's1',
+        id: 'c1',
         author: people.meera,
         timestamp: 'Yesterday',
-        body: 'Can someone bring extra gloves and water coolers? The park contact said the drinking fountain is unreliable.',
-        reactions: [{ emoji: '🧤', count: 4 }],
+        body: 'Packing list is finalized. Layers for evening, rain jacket just in case. No outside food in the retreat center kitchen.',
+        reactions: [{ emoji: '🎒', count: 11 }, { emoji: '🙏', count: 4 }],
+        replyCount: 5,
+        pinned: true,
+      },
+      {
+        id: 'c2',
+        author: people.vikram,
+        timestamp: 'Yesterday',
+        body: 'Reminder to fill out the waiver by midnight tonight — link is in the doc Meera shared.',
+        reactions: [{ emoji: '✅', count: 7 }],
         replyCount: 2,
       },
       {
-        id: 's2',
-        author: people.ravi,
-        timestamp: 'Yesterday',
-        body: 'I can cover gloves. Still need one more car from downtown if anyone is free.',
-        reactions: [{ emoji: '🙏', count: 7 }],
+        id: 'c3',
+        author: people.anand,
+        timestamp: '7:45 AM',
+        body: "Who's driving from downtown Chicago? Have two open seats if anyone needs a ride from the city.",
+        reactions: [{ emoji: '🚗', count: 5 }],
+        replyCount: 3,
+      },
+      {
+        id: 'c4',
+        author: people.rohit,
+        timestamp: '9:12 AM',
+        body: "Carpools are filling up fast — DM Anand if you're coming from downtown.",
+        reactions: [{ emoji: '👆', count: 3 }],
         replyCount: 1,
       },
     ],
@@ -393,32 +401,46 @@ export const centerBoards: CenterBoard[] = [
     subtitle: 'Verified CHYKs can coordinate rides, seva, and announcements here.',
     messages: [
       {
-        id: 'c1',
+        id: 'csj1',
         author: people.suresh,
         timestamp: '2h',
         body: 'Need two more volunteers for setup before Sunday satsang. Mostly chairs and AV. DM me if you can help.',
-        reactions: [
-          { emoji: '🙏', count: 8 },
-          { emoji: '🪔', count: 3 },
-        ],
+        reactions: [{ emoji: '🙏', count: 8 }, { emoji: '🪔', count: 3 }],
         replyCount: 5,
         pinned: true,
       },
+    ],
+  },
+  {
+    id: 'center-houston',
+    centerName: 'Chinmaya Houston',
+    title: 'Center board',
+    subtitle: 'Verified CHYKs can coordinate rides, seva, and announcements here.',
+    messages: [
       {
-        id: 'c2',
-        author: people.priya,
-        timestamp: 'Yesterday',
-        body: "Mother's Day prep — sign-up sheet is open for cooking, decoration, and reception desk.",
-        reactions: [{ emoji: '❤️', count: 6 }],
+        id: 'chou1',
+        author: people.vikram,
+        timestamp: '3h',
+        body: 'GCC State Finals are this Saturday at Chinmaya Mangalam — come cheer on the participants! 🙏',
+        reactions: [{ emoji: '🎉', count: 12 }, { emoji: '🙏', count: 6 }],
         replyCount: 4,
       },
+    ],
+  },
+  {
+    id: 'center-tri-state',
+    centerName: 'Chinmaya Mission Tri-State',
+    title: 'Center board',
+    subtitle: 'Verified CHYKs can coordinate rides, seva, and announcements here.',
+    messages: [
       {
-        id: 'c3',
-        author: people.karthik,
-        timestamp: 'Mon',
-        body: 'Uploading the chanting packet tonight so newer folks can print it ahead of class.',
-        reactions: [{ emoji: '📄', count: 5 }],
-        replyCount: 1,
+        id: 'cnj1',
+        author: people.rohit,
+        timestamp: '1h',
+        body: "33rd Aradhana Camp registration closes Friday — if you haven't signed up yet, now is the time.",
+        reactions: [{ emoji: '🏕️', count: 14 }, { emoji: '🙏', count: 9 }],
+        replyCount: 7,
+        pinned: true,
       },
     ],
   },
@@ -573,7 +595,8 @@ export function buildCenterBoard(input: {
   centerName: string
   subtitle: string
 }): CenterBoard {
-  const template = centerBoards[0]
+  const idx = input.id.split('').reduce((acc, c) => acc + c.charCodeAt(0), 0) % centerBoards.length
+  const template = centerBoards[idx]
   return {
     ...template,
     id: input.id,
