+++ packages/frontend/components/feed/index.ts
@@ -1,0 +1,7 @@
+export { FeedHeader } from './FeedHeader'
+export { NativeChatHeader } from './NativeChatHeader'
+export { FeedPostCard } from './FeedPostCard'
+export { FeedList } from './FeedList'
+export { PostThread } from './PostThread'
+export { FeedWorkspace } from './FeedWorkspace'
+export type { FeedPost, GroupBoard } from './types'
