+++ packages/frontend/components/feed/types.ts
@@ -1,0 +1,23 @@
+import type { BoardMessage, GroupKind } from '../connect'
+
+export type GroupBoard = {
+  id: string
+  kind: GroupKind
+  title: string
+  eyebrow: string
+  subtitle: string
+  meta: string
+  preview: string
+  unreadCount: number
+  messages: BoardMessage[]
+  distanceMi?: number
+}
+
+export type FeedPost = BoardMessage & {
+  groupId: string
+  groupKind: GroupKind
+  sourceTitle: string
+  sourceSubtitle: string
+  sourceLabel: string
+  replyMessages: BoardMessage[]
+}
